script_js() accepts SCRIPT_ATTACH_VERSION flag.
When used with SCRIPT_URL, Textpattern's version number
is attached to the link if the current install running is
a stable release.

E.g. 'textpattern.js' becomes 'textpattern.v4.6.js'.

Stable releases are detected based on the version number.
If the version contains a tag "-dev" the install is considered
as checkout of the dev branch.

This option is used to prevent clients from caching outdated
version of linked resources.

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@4321 2fea6a4d-a838-0410-917b-93a53c48e9d2