Prevent set_pref() from updating wrong rows.
set_pref() had a change to update different prefs row(s)
than the one the same statement inserted to the database.

This is due to having LIKE operator in update statement
while not sporting wildcards when checking existence or adding new prefs rows.

This means that wildcards would both end to the database
with the name, but also that update statement could update
different row(s) than intended.

This change removes that stranded LIKE operator which most likely was there by accident.

Closes issue #296

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@4367 2fea6a4d-a838-0410-917b-93a53c48e9d2