Selecting by label causes issues.
If the current value matches both a label and a value,
both of the options are selected.

For instance if a section's title is same as other's name,
both of the options are selected, leading into saving issues.
One of the options can not be used and needs section need
to be  re-selected each time an article is edited.

The label will never be sent to server in the first place. If the
array is not associative, numerical indexes are used for
values.

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@4610 2fea6a4d-a838-0410-917b-93a53c48e9d2