adminErrorHandler() doesn't terminate panel pages.
Terminating the script on error sounds bit dangerous and
extreme. That would kill any ongoing operation which
can lead to unexpected damage. We should ensure
that the error handlers don't do anything unexpected
or affect anything, but output if possible.

Since that change, I haven't seen anything but white pages
when developing plugins and that has made debugging
much more painful. I like many other, do typos.

The handler will still terminates the process on misc async
responses, since there is no better way of announcing the
error in 'text' response. Adding the error in the content
is not an option when the response is just one blob.

See issue #373

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@5427 2fea6a4d-a838-0410-917b-93a53c48e9d2