publicErrorHandler() doesn't display notices in testing mode.
There are number of Textpattern sites that leave their
production status to testing mode. This is due to whatever
reason, maybe due to it really never displaying errors.

But, because of this security feeling, users are
compromising their sites. PHP does not consistently sanitize
user-defined values present in notices, opening a
vector for XSS attacks.

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@5434 2fea6a4d-a838-0410-917b-93a53c48e9d2