Update txp:password_protect tag, and integrate with logins.
* If credentials aren't used, checks for Textpattern login.
* If checking login, additional privs can be provided.
* Can be used as a container tag.

Textpattern login is checked by the cookie. It does not
access login details through Basic, because of security and
replay attacks. Basic would persistently store your password
in the request, and allows no way of logging out.

The tag merely checks for existing login, instead of
authenticating in situations where actual login is checked.

When checking Textpattern login, the login attribute can
be used to restrict name access and privs can be used to
restrict privileges.

As a container tag, the tag hides the wrapped content if
the user doesn't have access to it. When using the Basic
authentication method, it prompts for login too, but doesn't
kill the page.

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@5590 2fea6a4d-a838-0410-917b-93a53c48e9d2