Add cancel button to Section Edit panel
Note that, unlike the other panels, this panel does not receive error conditions. All errors revert to the Section List panel. This is due to the way the panel is currently coded.

Also fixes Add Section functionality which erroneously imported the description from the default Section into every new Section.