This is bad - needs a TODO to remind us to improve it
Currently if the global comment pref is set to ‘off’, the section still
renders but empty of content, and with a clunky ‘empty’ classname in
order to hide it from screen.

This should be handled in a similar fashion to comment features
elsewhere in system - i.e. don’t render the region in HTML at all if
turned off globally.