No zero dates in SQL. Category name 64 chars. Re-create txp_token
MySQL 5.7 by default no longer allows 0000-00-00 dates, which were used as the default on several columns in TXP tables.
This patch removes those defaults for NOT NULL columns and sets the default to NULL for NULL columns (textpattern:Expires and txp_users:last_access)
Zero dates in existing rows are converted to sensible values that are allowed.

Categories, when created or edited can be 64 chars long. A few other tables that used these categories allowed for longer names. That doesn't make sense, because they can't be longer than 64 chars, so I've changed those colums to varchar(64).

txp_token was added to the update scripts, but without NOT NULL and it was also missing from txpsql.php. Added it there and recreated the table to be more in line with the other tables.