More consistent duplicate behaviour
Fixes issue #614. Changes are:

Duplicate from article write tab:
* reset posted time if that option is checked.
* always set last modified time to NOW.

Duplicate from articles list tab:
* don't change Posted and Expires dates.
* reset comments_count to zero.
* set last modified time to NOW (fixes issue #618).
* set status to draft if the old status was live/sticky

One difference remains. On the write tab, the url_title is regenerated from the title when duplicating. On the articles list tab, this doesn't happen.

Side effect: for new articles, the last modified time is now always set to NOW(). Previously, for articles set in the future it was set to the posted date. However when editing such an article, it would be reset to NOW(), so that didn't make much sense.