Re-introduce tag builder on Pages panel
Mostly works, a few bells and whistles to add, and there are no doubt plenty of JS optimisations to be done.

One known bug at present: after hitting 'Build', it populates the textarea, but hitting 'Build' a second time renders the tag builder in place of the entire page. This is because the 'Build' button isn't properly delegated so it loses its event handler after first submission and thus does a regular POST.