Refactor label tools in inputLabel()
Instead of introducing a new argument to inputLabel(), retrofit the $label
parameter to be able to take an array, the second argument of which
may contain the HTML block to append to the label.

The label can still be a string, or the first array arg may be empty/null,
in which case $name will still be used for the plain label. The passed
label text string is also still passed through gTxt().