Setup no longer uses a fixed list of langs
The available languages are now read from the filesystem. This has the advantage that we can now more easily render the setup routine in RTL too, as is done in this commit.

There's probably a more elegant way of getting the current language file on each page of the wizard than iterating over the directory and finding it, but this is just working with what we have available. Pre-filling every language's metadata in the SESSION is possible, but addcslashes() complains about array-based session variables, so that portion of the code would need changing to include an exception to the session variable iterator.

RPC language provision has also been removed from the setup routine.