Ahem, yeah, fix language panel... sort of
Botched logic. Still doesn't install the plugin strings for a newly installed language, even though the installed_langs array is updated. Needs investigation.

Fixes the multi_query slightly too, so it returns correct results for INSERTS. In theory it should return a series of result sets but is untested as yet for SELECTs.