Permit custom fields to be named '0'
Although it's less than ideal, it's easier than trying to trap custom field values on the Preferences panel. And, arguably, it's user content so we shouldn't retsrict the CF names, despite how strange it may appear to use a field with a name of '0'. Going forward, we'll need to respect this when unlimited CFs become reality.

Fixes issue #1080. Thanks, petecooper.