Verify all new plugin interactions
This is phase one of plugin handling updates. See #1489.

Drag/drop either .txt, .php or .zip onto the file browser in the Plugins panel to upload, or paste .txt file contents into the upload box as normal. All such actions will display the Verify step. No files are committed to the textpattern/plugins directory unless you proceed. Cancelling the process shows an alert and tidies up any temporary files..

The verify step now also displays any additional data that may be bundled inside the plugin (things destined for the 'data' column in the table). For .zip packages, a list of files that are going to be unpacked is displayed in an 'Upload' textarea. This might be a temporary measure. We may elect to permit checkboxes alongside these files so administrators can choose what is uploaded.

Plugins that have been previously uploaded/installed are presumed safe enough to simply relink, so these do _not_ go through the verification step. Same with refresh from disk: it is immediate.