Permit plugin upload from URL
Drag/paste a valid plugin URL of a .txt, .php or .zip file into the 'Install plugin' textarea and click Upload. The URL contents will be fetched and verification will be requested as normal.

Note this does allow you to install absolute gibberish from any URL on the web (even HTML) but the same could be said for our ability to upload a "PHP" file, which could really be a cake recipe with a .php extension. The addition of the enforced verification, however, makes it far harder to sneak any crap past an administrator.

We could do some rudimentary checks for 'acceptableness' of content if we deem it necessary. e.g. if we're passed a PHP file, does it appear to be PHP/a Txp plugin template.