Permit database creation during setup
A new select option allows you to immediately (as soon as you click Next) create a database instead of having to manually do it outside Textpattern. The database will only be created if it doesn't already exist; you can't overwrite one.

As part of this change, the database user account permissions are verified against a required list of privileges (aka grants). If the account doesn't meet the minimum spec, installation is aborted and you are informed which permissions you require.

This system extends to automated installers too. If you wish to auto-create the given database in your JSON config file, add ["database"]["create"] = "1".