Add plugin lifecycle events 'upgraded' and 'downgraded'
Handy for specific situations like upgrading table data when migrating from version to version. Note that semver is adhered to.

Two variables are transmitted along for the ride in the third parameter passed to the lifecycle callback: 'previous' and 'version'. This allows you to conditionally test either value to perform stepped migrations.