Tentative reworking of the File Edit panel
Layout more like the Images panel, with additional functionality:
* Delete file button
* Extra metadata shown (file size, last modified)
* Rename file
* Edit download count

The layout is a little shonky. Would be better if we could somehow offer a preview (thumbnail?) or icon that represents the file/type, but that might be process intensive. May revert to a single central block, but with file replace in its own control panel area (because it needs to have its own form).

Ideas stolen shamelessly from the adi_file_tab plugin. Thank you, Adi. There may be other features that can be transplanted.